// src/app/modules/shared/models/user.ts
export class User {
  iat: string;
  user: {
    Address: string;
    created_at: string;
    email_Id: string;
    firstname: any;
    imagePath: any;
    lastname: any;
    mobileNumber: string;
    password: string;
    roles: string;
    signup_type: any;
    status: any;
    updated_at: string;
    user_desc: string;
    vendorId: string;
  };
}